+++ src/components/Header/CartTextButton/CartTextButton.tsx
@@ -1,7 +1,7 @@
 import { useRecoilValue } from 'recoil';
 import { styled } from 'styled-components';
-import { cartSelector } from '../../store/cart';
-import { WIDTH } from '../../styles/mediaQuery';
+import { cartSelector } from '../../../store/cart';
+import { WIDTH } from '../../../styles/mediaQuery';
 
 const CartTextButton = () => {
   const { cartsQuantity } = useRecoilValue(cartSelector);
