+++ src/components/ProductCardList/ProductCardList.tsx
@@ -2,7 +2,7 @@ import { useRecoilValue } from 'recoil';
 import { styled } from 'styled-components';
 import { fetchedProductListSelector } from '../../store/asyncSelector';
 import { WIDTH } from '../../styles/mediaQuery';
-import ProductCard from '../ProductCard/ProductCard';
+import ProductCard from './ProductCard/ProductCard';
 import { cartAtom } from '../../store/cart';
 
 const ProductCardList = () => {
